Rename RayCluster folder to Ray since the group is Ray
