+++ ray-operator/apis/ray/v1alpha1/register.go
