+++ ray-operator/controllers/ray/common/rbac.go
@@ -1,8 +1,8 @@
 package common
 
 import (
-	"github.com/ray-project/kuberay/ray-operator/apis/raycluster/v1alpha1"
-	"github.com/ray-project/kuberay/ray-operator/controllers/raycluster/utils"
+	"github.com/ray-project/kuberay/ray-operator/apis/ray/v1alpha1"
+	"github.com/ray-project/kuberay/ray-operator/controllers/ray/utils"
 	v1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
